+++ protected/humhub/modules/comment/widgets/Comment.php
@@ -1,42 +1,52 @@
 <?php
 
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ */
+
 namespace humhub\modules\comment\widgets;
+
+use yii\helpers\Url;
 
 /**
  * This widget is used to show a single comment.
- *
  * It will used by the CommentsWidget and the CommentController to show comments.
  *
  * @package humhub.modules_core.comment
- * @since 0.5
  */
 class Comment extends \yii\base\Widget
 {
 
     /**
-     * @var Comment object to display
+     * @var \humhub\modules\comment\models\Comment the comment
      */
     public $comment = null;
 
     /**
-     * Indicates the comment was just edited
-     *
-     * @var boolean
+     * @var boolean indicator that comment has just changed
      */
     public $justEdited = false;
 
     /**
-     * Executes the widget.
+     * @inheritdoc
      */
     public function run()
     {
-
-        $user = $this->comment->user;
+        $deleteUrl = Url::to(['/comment/comment/delete', 'contentModel' => $this->comment->object_model, 'contentId' => $this->comment->object_id, 'id' => $this->comment->id]);
+        $editUrl = Url::to(['/comment/comment/edit', 'contentModel' => $this->comment->object_model, 'contentId' => $this->comment->object_id, 'id' => $this->comment->id]);
+        $loadUrl = Url::to(['/comment/comment/load', 'contentModel' => $this->comment->object_model, 'contentId' => $this->comment->object_id, 'id' => $this->comment->id]);
 
         return $this->render('showComment', [
-            'comment' => $this->comment,
-            'user' => $user,
-            'justEdited' => $this->justEdited,
+                    'comment' => $this->comment,
+                    'user' => $this->comment->user,
+                    'justEdited' => $this->justEdited,
+                    'deleteUrl' => $deleteUrl,
+                    'editUrl' => $editUrl,
+                    'loadUrl' => $loadUrl,
+                    'canWrite' => $this->comment->canWrite(),
+                    'canDelete' => $this->comment->canDelete(),
         ]);
     }